Create Update_Map.ipynb
Exploring options for automatically updating the regions map based on EHP. Currently the fill color is based on the first place team and the outline is based on the second place team. Transparency is being scaled so the more transparent the fill is, the closer the second place team is to the first.